+++ src/cfl-converter.ts
@@ -1,30 +1,28 @@
 #!/usr/bin/env node
-import chalk from 'chalk';
-import program from 'commander';
-import download from 'download';
 import fs from 'fs';
-import JSZip from 'jszip';
-import lzma from 'lzma-purejs';
 
-import { BinaryReader, BinaryWriter, Encoding  } from 'csharp-binary-stream';
-
-interface CFLEntry {
-  compression: number;
-  offset: number;
-  length: number;
-  size: number;
-
-  name: string;
-  hash: string | null;
-
-  contents: Uint8Array;
-}
+import { convert } from '.';
 
 async function main() {
+  let program;
+  let download;
+
+  try {
+    program = (await import('commander')).default;
+    download = (await import('download')).default;
+  } catch {
+    console.error('One or more optional dependencies is missing! Please install them first.');
+    process.exit(1);
+  }
+
   program
   .version('0.0.1')
-  .option('-I, --input <files>', '.CFL file(s) to convert.', (value: string): string[] => value.split(','))
-  .option('-P, --products <ids>', 'List of Product IDs to retrieve and convert.', (value: string) => value.split(','));
+  .option('-I, --input <files>', '.CFL file(s) to convert.', (value: string): string[] => value.split(','), [])
+  .option('-P, --products <ids>', 'List of Product IDs to retrieve and convert.', (value: string) => value.split(','), [])
+  .option('-Q, --quiet', 'Stop the program from logging progress.', Boolean, false)
+  .option('-D, --dry-run', 'Skip outputting to a file and just attempt to convert.', Boolean, false)
+  .option('--ignore-missing', 'Ignore IDs that point to non-existant products.', Boolean, false)
+  .option('--ignore-extension', 'Attempt to process all files regardless of extension.', Boolean, false);
 
   program.on('--help', () => {
     console.log('');
@@ -35,10 +33,10 @@ async function main() {
 
   program.parse(process.argv);
 
-  const products: Buffer[] = (await Promise.all((program.products || []).map(async (product: string): Promise<Buffer | null> =>  {
+  async function downloadProduct(id: number | string): Promise<ArrayBufferLike | null> {
     for (let i = 100; i > 0; i--) {
       try {
-        const file = await download(`http://userimages-akm.imvu.com/productdata/${product.trim()}/${i}`);
+        const file = await download(`http://userimages-akm.imvu.com/productdata/${typeof id === 'number' ? id : id.trim()}/${i}`);
 
         return Buffer.from(file);
       } catch {
@@ -47,99 +45,72 @@ async function main() {
     }
 
     return null;
-  })) as Array<Buffer | null>).filter((value: Buffer | null) => value != null);
+  }
 
-  const inputs: Array<string | Buffer> = [...program.input || [], ...products];
+  const options = {
+    input: program.input as string[],
+    log: !program.quiet,
+    products: program.products as string[],
+  };
 
-  console.log(chalk.yellowBright(`Beginning processing on ${inputs.length} file${inputs.length > 1 ? 's' : ''}. . .\n`));
+  if (options.products.length > 0 && options.log) {
+    console.log(`Downloading ${options.products.length} CFL files...\n`);
+  }
 
-  let count = 0;
-  for (const input of inputs) {
-    if (typeof input === 'string' && !input.toLocaleLowerCase().endsWith('.cfl')) {
-      console.log(chalk.redBright('This program only supports files ending with the .CFL extension!'));
+  const products = await Promise.all(options.products.map(downloadProduct));
+
+  if (!program.ignoreMissing) {
+    products.forEach((value) => {
+      if (value == null) {
+        console.error('One or more CFLs failed to download! (Use the --ignore-missing flag to override)');
+      }
+    });
+  }
+
+  if (!program.ignoreExtension) {
+    options.input.forEach((value) => {
+      if (!value.toLocaleLowerCase().endsWith('.cfl')) {
+        console.error('This program only supports files with the .CFL extension! (Use the --ignore-extension flag to override)');
+        process.exit(1);
+      }
+    });
+  }
+
+  const inputs = [...options.input.map((input) => fs.readFileSync(input.trim())), ...products];
+  const names = [...options.input.map((input) => input.toLocaleLowerCase().endsWith('.cfl') ? input.slice(0, -4) : input), ...options.products];
+
+  if (options.log) {
+    console.log(`Beginning processing on ${inputs.length} file(s). . .\n`);
+  }
+
+  await Promise.all(inputs.map(async (input, index) => {
+    if (input == null) {
       return;
     }
 
-    console.log(chalk.blueBright(`Processing '${typeof input === 'string' ? input : 'download'}'. . .`));
-
-    const reader = new BinaryReader(typeof input === 'string' ? fs.readFileSync(input.trim()) : input);
-
-    // CFL Header
-    // This information is exclusive to the CFL format
-
-    const hashed = reader.readChars(4, Encoding.Utf8) === 'DFL3';
-
-    reader.position = reader.readUnsignedInt();
-
-    const format = reader.readInt();
-
-    console.log(chalk.greenBright('\tDecoded CFL header.'));
-
-    // Decompress the LZMA section of the CFL file
-    const content = Uint8Array.from(reader.readBytes(reader.readInt()));
-    const data = new BinaryReader(format === 4 ? decompress(content) : content);
-
-    console.log(chalk.greenBright('\tDecompressed CFL content.'));
-
-    // Parse the CFL entries
-    const entries: CFLEntry[] = [];
-
-    let position = 0;
-    while (data.length > position) {
-      // CFL entry header data
-      const size = data.readInt();
-      const offset = data.readInt();
-      const compression = data.readInt();
-      const name = data.readChars(data.readShort(), Encoding.Utf8);
-
-      reader.position = offset;
-
-      // Decompress the LZMA section of the entry
-      const entry = Uint8Array.from(reader.readBytes(reader.readUnsignedInt()));
-      const contents = compression === 4 ? decompress(entry) : entry;
-
-      // Calculate hash and entry size
-      const hash: string | null = hashed ? data.readChars(data.readInt(), Encoding.Utf8) : null;
-      const length = 14 + name.length + (hash !== null && hash.match(/^\s*$/) === null ? 4 + hash.length : 0);
-
-      entries.push({
-        compression,
-        contents,
-        hash,
-        length,
-        name,
-        offset,
-        size,
-      });
-
-      position += length;
+    if (options.log) {
+      console.log(`Processing file ${names[index]}. . .`);
     }
 
-    console.log(chalk.greenBright(`\tParsed ${entries.length} CFL ${entries.length !== 1 ? 'entries' : 'entry'}.`));
+    try {
+      const file = await convert(input, { log: options.log });
 
-    count++;
+      if (!program.dryRun) {
+        fs.writeFileSync(`${names[index]}.chkn`, file);
+        if (options.log) {
+          console.log(`Wrote to ${names[index]}.chkn.\n`);
+        }
+      } else if (options.log) {
+        console.log(`Successfully converted ${names[index]}.\n`);
+      }
+    } catch {
+      console.log(`Failed processing ${names[index]}!\n`);
+    }
+  }));
 
-    const zip = new JSZip();
-
-    entries.forEach((entry: CFLEntry) => {
-      zip.file(entry.name, entry.contents);
-    });
-
-    zip.generateNodeStream({ type: 'nodebuffer', streamFiles: true })
-      .pipe(fs.createWriteStream(`${typeof input === 'string' ? input.slice(0, -4) : count}.chkn`))
-      .on('close', () => console.log(chalk.greenBright(`\tWrote to ${typeof input === 'string' ? input.slice(0, -4) : count}.chkn.\n`)));
+  if (options.log) {
+    console.log(`Finished processing ${inputs.length} file(s)!`);
   }
-
-  console.log(`Finished processing ${inputs.length} file${inputs.length > 1 ? 's' : ''}!`);
-}
-
-function decompress(data: Uint8Array | Buffer): Uint8Array {
-  const reader = new BinaryReader(data);
-  const writer = new BinaryWriter();
-  // Read the first 5 bytes, they are the properties
-  // Set size to -1 to just read to EOS
-  lzma.decompress(reader.readBytes(5), reader, writer, -1);
-  return writer.toUint8Array();
 }
 
 main();